Modified User Class to add some attributes and validate mail using javax validator dependency
