--- src/main/java/gov/iti/jets/Admin/AdminController.java
@@ -1,5 +1,0 @@
-package gov.iti.jets.Admin;
-
-public class AdminController {
-    
-}
