--- src/main/java/gov/iti/jets/App/App.java
@@ -1,7 +1,0 @@
-package gov.iti.jets.App;
-
-public class App {
-    public static void main(String[] args) {
-        System.out.println("Hello, World!");        
-    }
-}
