--- src/main/java/gov/iti/jets/Auth/AuthFilter.java
@@ -1,5 +1,0 @@
-package gov.iti.jets.Auth;
-
-public class AuthFilter {
-    
-}
