--- src/main/java/gov/iti/jets/Cart/Cart.java
@@ -1,5 +1,0 @@
-package gov.iti.jets.Cart;
-
-public class Cart {
-    
-}
