--- src/main/java/gov/iti/jets/Common/Role.java
@@ -1,5 +1,0 @@
-package gov.iti.jets.Common;
-
-public class Role {
-    
-}
