--- src/main/java/gov/iti/jets/GenericDao/GenericDAO.java
@@ -1,13 +1,0 @@
-package gov.iti.jets.GenericDao;
-
-import java.util.Set;
-
-public interface GenericDAO<T> {
-    T save(T entity);
-  
-    Set<T> findAll();
-
-    T findById(java.lang.Integer id);
-
-    void delete(Integer id);
-}
