--- src/main/java/gov/iti/jets/Order/Order.java
@@ -1,5 +1,0 @@
-package gov.iti.jets.Order;
-
-public class Order{
-    
-}
