--- src/main/java/gov/iti/jets/Product/ProductService.java
@@ -1,5 +1,0 @@
-package gov.iti.jets.Product;
-
-public class ProductService {
-    
-}
