--- src/main/java/gov/iti/jets/User/UserController.java
@@ -1,5 +1,0 @@
-package gov.iti.jets.User;
-
-public class UserController {
-    
-}
