+++ src/main/java/gov/iti/jets/admin/ProductManagement.java
@@ -1,0 +1,5 @@
+package gov.iti.jets.admin;
+
+public class ProductManagement {
+    
+}
