+++ src/main/java/gov/iti/jets/app/App.java
@@ -1,0 +1,7 @@
+package gov.iti.jets.app;
+
+public class App {
+    public static void main(String[] args) {
+        System.out.println("Hello, World!");        
+    }
+}
