+++ src/main/java/gov/iti/jets/auth/AuthFilter.java
@@ -1,0 +1,5 @@
+package gov.iti.jets.auth;
+
+public class AuthFilter {
+    
+}
