+++ src/main/java/gov/iti/jets/common/Address.java
@@ -1,0 +1,17 @@
+package gov.iti.jets.common;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Address {
+    private String street;
+    private String city;
+
+
+}
