+++ src/main/java/gov/iti/jets/common/HibernateUtil.java
@@ -1,0 +1,5 @@
+package gov.iti.jets.common;
+
+public class HibernateUtil {
+    
+}
