+++ src/main/java/gov/iti/jets/common/UserRole.java
@@ -1,0 +1,6 @@
+package gov.iti.jets.common;
+
+public enum UserRole {
+    ADMIN,
+    CUSTOMER
+}
