+++ src/main/java/gov/iti/jets/order/Order.java
@@ -1,0 +1,5 @@
+package gov.iti.jets.order;
+
+public class Order{
+    
+}
