+++ src/main/java/gov/iti/jets/persistence/CustomPersistenceUnit.java
@@ -1,0 +1,115 @@
+package gov.iti.jets.persistence;
+
+import com.zaxxer.hikari.HikariDataSource;
+import jakarta.persistence.SharedCacheMode;
+import jakarta.persistence.ValidationMode;
+import jakarta.persistence.spi.ClassTransformer;
+import jakarta.persistence.spi.PersistenceUnitInfo;
+import jakarta.persistence.spi.PersistenceUnitTransactionType;
+
+import javax.sql.DataSource;
+import java.net.URL;
+import java.util.List;
+import java.util.Properties;
+
+public class CustomPersistenceUnit implements PersistenceUnitInfo {
+    @Override
+    public String getPersistenceUnitName() {
+        return "mysql";
+    }
+
+    @Override
+    public String getPersistenceProviderClassName() {
+        return "org.hibernate.jpa.HibernatePersistenceProvider";
+    }
+
+    @Override
+    public PersistenceUnitTransactionType getTransactionType() {
+        return PersistenceUnitTransactionType.RESOURCE_LOCAL;
+    }
+
+    @Override
+    public DataSource getJtaDataSource() {
+
+
+        try {
+            HikariDataSource dataSource = new HikariDataSource();
+            dataSource.setJdbcUrl("jdbc:mysql://localhost:3306/hibernate_jpa");
+            dataSource.setUsername("ghandy");
+            dataSource.setPassword("ghandy");
+            return dataSource;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+        } finally {
+
+        }
+        return null;
+    }
+
+    @Override
+    public DataSource getNonJtaDataSource() {
+       return null;
+    }
+
+
+
+    @Override
+    public List<String> getMappingFileNames() {
+        return List.of();
+    }
+
+    @Override
+    public List<URL> getJarFileUrls() {
+        return List.of();
+    }
+
+    @Override
+    public URL getPersistenceUnitRootUrl() {
+        return null;
+    }
+
+    @Override
+    public List<String> getManagedClassNames() {
+        return List.of();
+    }
+
+    @Override
+    public boolean excludeUnlistedClasses() {
+        return false;
+    }
+
+    @Override
+    public SharedCacheMode getSharedCacheMode() {
+        return null;
+    }
+
+    @Override
+    public ValidationMode getValidationMode() {
+        return null;
+    }
+
+    @Override
+    public Properties getProperties() {
+        return null;
+    }
+
+    @Override
+    public String getPersistenceXMLSchemaVersion() {
+        return "";
+    }
+
+    @Override
+    public ClassLoader getClassLoader() {
+        return null;
+    }
+
+    @Override
+    public void addTransformer(ClassTransformer classTransformer) {
+
+    }
+
+    @Override
+    public ClassLoader getNewTempClassLoader() {
+        return null;
+    }
+}
