+++ src/main/java/gov/iti/jets/product/ProductService.java
@@ -1,0 +1,5 @@
+package gov.iti.jets.product;
+
+public class ProductService {
+    
+}
