+++ src/main/java/gov/iti/jets/user/UserController.java
@@ -1,0 +1,5 @@
+package gov.iti.jets.user;
+
+public class UserController {
+    
+}
