+++ src/main/java/gov/iti/jets/user/UserRepository.java
@@ -1,0 +1,27 @@
+package gov.iti.jets.user;
+
+import gov.iti.jets.genericDao.GenericDaoImpl;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.CriteriaQuery;
+import jakarta.persistence.criteria.Root;
+
+public class UserRepository extends GenericDaoImpl<User> {
+
+    public UserRepository() {
+        super(User.class);
+    }
+
+    public User getUserByUsername(String username) {
+        CriteriaBuilder cb = em.getCriteriaBuilder();
+
+        CriteriaQuery<User> q = cb.createQuery(User.class);
+
+        Root<User> user = q.from(User.class);
+
+        q.where(cb.equal(user.get("username"), username));
+
+        q.select(user).distinct(true);
+
+        return em.createQuery(q).getSingleResult();
+    }
+}
