+++ src/main/java/gov/iti/jets/user/UserService.java
@@ -1,0 +1,12 @@
+package gov.iti.jets.user;
+
+public class UserService {
+    UserRepository userRepository;
+    public UserService() {
+        userRepository = new UserRepository();
+    }
+
+    public User getUserByUsername(String username) {
+        return userRepository.getUserByUsername(username);
+    }
+}
