+++ src/main/java/gov/iti/jets/user/UserValidation.java
@@ -1,0 +1,5 @@
+package gov.iti.jets.user;
+
+public class UserValidation {
+    
+}
